Decomposed classes into simpler structure
