+++ src/main/java/cashpiles/model/PeriodsList.java
@@ -1,0 +1,7 @@
+package cashpiles.model;
+
+import java.util.ArrayList;
+
+@SuppressWarnings("serial")
+class PeriodsList extends ArrayList<BudgetPeriod> implements PeriodsView {
+}
